+++ ruoyi-common/src/main/java/com/ruoyi/common/utils/ip/AddressUtils.java
@@ -26,7 +26,7 @@ public class AddressUtils
 
     public static String getRealAddressByIP(String ip)
     {
-        // 内网不查询
+        // 内网地址不查询
         if (IpUtils.internalIp(ip))
         {
             return "内网IP";
@@ -35,7 +35,7 @@ public class AddressUtils
         {
             try
             {
-                String rspStr = HttpUtils.sendGet(IP_URL, "ip=" + ip + "&json=true", Constants.GBK);
+                String rspStr = HttpUtils.sendGet(IP_URL, "ip=" + ip, Constants.GBK);
                 if (StringUtils.isEmpty(rspStr))
                 {
                     log.error("获取地理位置异常 {}", ip);
@@ -44,7 +44,11 @@ public class AddressUtils
                 JSONObject obj = JSON.parseObject(rspStr);
                 String region = obj.getString("pro");
                 String city = obj.getString("city");
-                return String.format("%s %s", region, city);
+                if (StringUtils.isEmpty(region) || StringUtils.isEmpty(city))
+                {
+                    return UNKNOWN;
+                }
+                return region + " " + city;
             }
             catch (Exception e)
             {